+++ src/types/Notification.ts
@@ -1,0 +1,5 @@
+export interface INotification {
+  id: number;
+  message: string;
+  type: "info" | "success" | "warning" | "error";
+}
